+++ Tests/SentryTests/SentrySessionTests.swift
@@ -92,5 +92,22 @@ class SentrySessionTestsSwift: XCTestCase {
         XCTAssertEqual(expected.errors, session.errors)
         XCTAssertNil(session.releaseName)
     }
+    
+    func testInitWithJson_IfJsonContainsWrongValues_DefaultValuesAreUsed() {
+        let expected = SentrySession(releaseName: "release")
+        var serialized = expected.serialize()
+        serialized["sid"] = ""
+        serialized["started"] = "20"
+        serialized["status"] = "20"
+        serialized["timestamp"] = "20"
 
+        let session = SentrySession(jsonObject: serialized)
+        
+        XCTAssertNotEqual(expected.sessionId, session.sessionId)
+        XCTAssertEqual(expected.started, session.started)
+        XCTAssertEqual(expected.status, session.status)
+        XCTAssertEqual(expected.sequence, session.sequence)
+        XCTAssertEqual(expected.errors, session.errors)
+        XCTAssertNil(session.timestamp)
+    }
 }